Switched UI layout manager
Switched from using the pack manager to the grid manager in an effort to fix bugs where certain ui elements would disappear when large images were displayed.  Images will most likely not properly scale for multi monitor set ups but that is a battle for another day.

Also added the ability to press the enter key instead of the search button in the search menu.